requirements.txt
requests
python-dotenv
pytelegrambotapi